Extract form permissions tests
